feat: enhance action and auth
