+++ packages/pieces/community/stripe/src/lib/actions/create-price.ts
@@ -3,7 +3,7 @@ import { httpClient, HttpMethod } from '@activepieces/pieces-common';
 import { stripeAuth } from '../..';
 import { stripeCommon } from '../common';
 // CHANGED: Import the specific dropdown you need
-import { productIdDropdown } from '../common/props';
+import { productIdDropdown } from '../common';
 
 export const stripeCreatePrice = createAction({
     name: 'create_price',
