+++ packages/pieces/community/stripe/src/lib/actions/find-invoice.ts
@@ -3,7 +3,7 @@ import { httpClient, HttpMethod } from '@activepieces/pieces-common';
 import { stripeAuth } from '../..';
 import { stripeCommon } from '../common';
 // CHANGED: Import the specific dropdown you need
-import { invoiceIdDropdown } from '../common/props';
+import { invoiceIdDropdown } from '../common';
 
 export const stripeFindInvoice = createAction({
     name: 'find_invoice',
