+++ packages/pieces/community/stripe/src/lib/actions/retrieve-payout.ts
@@ -3,7 +3,7 @@ import { httpClient, HttpMethod } from '@activepieces/pieces-common';
 import { stripeAuth } from '../..';
 import { stripeCommon } from '../common';
 // CHANGED: Import the specific dropdown you need
-import { payoutIdDropdown } from '../common/props';
+import { payoutIdDropdown } from '../common';
 
 export const stripeRetrievePayout = createAction({
     name: 'retrieve_payout',
