+++ packages/pieces/community/stripe/src/lib/actions/update-customer.ts
@@ -3,7 +3,7 @@ import { httpClient, HttpMethod } from '@activepieces/pieces-common';
 import { stripeAuth } from '../..';
 import { stripeCommon } from '../common';
 // CHANGED: Import the specific dropdown you need
-import { customerIdDropdown } from '../common/props';
+import { customerIdDropdown } from '../common';
 
 export const stripeUpdateCustomer = createAction({
     name: 'update_customer',
